refactor typings
