+++ src/rpc/typings/rpc.d.ts
@@ -1,0 +1,11 @@
+export interface RPCRequest {
+  method: string,
+  params: Array<any>,
+  id: number
+}
+
+export interface RPCResponse {
+  jsonrpc: string,
+  id: number,
+  result: any
+}
