+++ src/wallet/typings/Balance.d.ts
@@ -1,0 +1,22 @@
+import { Fixed8 } from '../../utils';
+import { Transaction } from '../../transactions/index'
+import { AssetBalance } from './AssetBalance';
+
+export class Balance {
+  constructor(bal?: Balance)
+
+  address: string
+  net: 'MainNet' | 'TestNet'
+  assetSymbols: string[]
+  assets: { [index: string]: AssetBalance }
+  tokenSymbols: string[]
+  tokens: { [index: string]: number }
+
+  static import(jsonString: string): Balance
+
+  addAsset(sym: string, assetBalance?: AssetBalance): this
+  addToken(sym: string, tokenBalance?: number | Fixed8): this
+  applyTx(tx: Transaction, confirmed?: boolean): this
+  export(): string
+  verifyAssets(url: string): Promise<Balance>
+}
