+++ src/wallet/typings/ClaimItem.d.ts
@@ -1,0 +1,14 @@
+import { Fixed8 } from '../../utils';
+
+export function ClaimItem (config: ClaimItemObj): ClaimItemObj;
+
+export function exportClaimItem(claimItem: ClaimItemObj): ClaimItemObj;
+
+export interface ClaimItemObj {
+  claim?: number;
+  txid?: string;
+  index?: number;
+  value?: number;
+  start?: number;
+  end?: number;
+}
