+++ src/wallet/typings/nep2.d.ts
@@ -1,0 +1,6 @@
+import { ScryptParams } from './core';
+
+export function encrypt(wifKey: string, keyphrase: string, scryptParams?: ScryptParams): string
+export function decrypt(encryptedKey: string, keyphrase: string, scryptParams?: ScryptParams): string
+export function encryptAsync(wifKey: string, keyphrase: string, scryptParams?: ScryptParams): string
+export function decryptAsync(encryptedKey: string, keyphrase: string, scryptParams?: ScryptParams): string
