+++ src/wallet/typings/verify.d.ts
@@ -1,0 +1,5 @@
+export function isNEP2(nep2: string): boolean
+export function isWIF(wif: string): boolean
+export function isPrivateKey(key: string): boolean
+export function isPublicKey(key: string, encoded?: boolean): boolean
+export function isAddress(address: string): boolean
